as-call-service/requirements.txt
# Core FastAPI and Python dependencies
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database dependencies
asyncpg>=0.29.0
sqlalchemy>=2.0.0
alembic>=1.13.0

# HTTP client for service communication
httpx>=0.25.2

# Validation and utilities
email-validator>=2.1.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6

# Testing dependencies
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Development dependencies
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.0

# Environment and configuration
python-dotenv>=1.0.0